chore: rename selfDependant to selfDependent
